+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Changes
 * [Documentation]: Updated Getting Started docs to include IRIDA log in info and an updated install-libs.sh run command.
 * [Developer/UI]: Upgraded to FastQC 0.11.9.
 * [UI]: Added FAST5 format support.  FAST5 support is in beta and is subject to change.  See more in our upgrade notes at <https://irida.corefacility.ca/documentation/administrator/upgrades/#2005>.
+* [UI]: Updated user groups to use Ant Design.
 
 19.09 to 20.01
 --------------